Move add to flow json with add contract to flowkit
