+++ flowkit/flowkit.go
@@ -23,6 +23,7 @@ import (
 	"context"
 	"crypto/rand"
 	"fmt"
+	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -254,6 +255,12 @@ func UpdateExistingContract(updateExisting bool) UpdateContract {
 	}
 }
 
+func getFilenameFromPath(path string) string {
+	filename := filepath.Base(path)
+	extension := filepath.Ext(filename)
+	return filename[0 : len(filename)-len(extension)]
+}
+
 // AddContract to the Flow account provided and return the transaction ID.
 //
 // If the contract already exists on the account the operation will fail and error will be returned.
@@ -363,6 +370,24 @@ func (f *Flowkit) AddContract(
 	}
 	if trx.Error != nil {
 		return tx.FlowTransaction().ID(), false, trx.Error
+	}
+
+	// If success, add the new account to the state
+	filename := getFilenameFromPath(contract.Location)
+	d := state.Deployments().ByAccountAndNetwork(account.Name, f.network.Name)
+	if d != nil {
+		d.AddContract(config.ContractDeployment{
+			Name: getFilenameFromPath(filename),
+		})
+	}
+	state.Contracts().AddOrUpdate(config.Contract{
+		Name:     filename,
+		Location: contract.Location,
+	})
+
+	err = state.SaveDefault()
+	if err != nil {
+		return tx.FlowTransaction().ID(), false, err
 	}
 
 	return sentTx.ID(), updateExisting, err
